Consolidate firebase-app logic in firebase-module
